encapsulation, switch statements
